backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
google-generativeai==0.3.2
gunicorn==21.2.0
